% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{abort_bad_argument}
\alias{abort_bad_argument}
\title{Abord bad argument}
\usage{
abort_bad_argument(arg1, must, not = NULL, arg2 = NULL, same = NULL)
}
\arguments{
\item{arg1}{An argument}

\item{must}{What arg must be}

\item{not}{Optional. What arg must not be.}

\item{arg2}{Another argument}

\item{same}{The same type as arg1}
}
\value{
A stop statement
}
\description{
Abord bad argument
}
