% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_report.R
\name{analyse_kpis}
\alias{analyse_kpis}
\title{Analyse KPIs}
\usage{
analyse_kpis(data, logic_model)
}
\arguments{
\item{data}{A data.frame with actual KPI values}

\item{logic_model}{A logic model list with target KPIs}
}
\value{
A data.frame with KPI analysis
}
\description{
Compares actual vs. target KPI values and calculates percentage achieved and performance gaps.
}
