% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{bind_others}
\alias{bind_others}
\title{Bind other and parent values}
\usage{
bind_others(.tbl, other, id_col)
}
\arguments{
\item{.tbl}{A tibble of data}

\item{other}{A character vector of the start of all other column names. E.g., other = "other_"}

\item{id_col}{Survey id, usually uuid (quoted or unquoted)}
}
\value{
A tibble with other and parent other questions and values
}
\description{
Bind other and parent values
}
\details{
Made to ease the use of other_cols and other_parent_cols
}
