% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{buffer_bbox}
\alias{buffer_bbox}
\title{Bbbox buffer}
\usage{
buffer_bbox(sf_obj, buffer = 0)
}
\arguments{
\item{sf_obj}{A `sf` object}

\item{buffer}{A buffer, either one value or a vector of 4 values (left, bottom, right, top). Default to 0.}
}
\value{
A bbox with a buffer
}
\description{
Bbbox buffer
}
