% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_analysis}
\alias{check_analysis}
\title{Check analysis}
\usage{
check_analysis(design, survey, choices, analysis, col, group = NULL, level)
}
\arguments{
\item{design}{A \code{srvyr} design}

\item{survey}{The survey sheet}

\item{choices}{The choices sheet}

\item{analysis}{Some analysis}

\item{col}{A variable from design}

\item{group}{A grouping variable, quoted}

\item{level}{Some confidence level}
}
\description{
Check analysis
}
\seealso{
Other functions to check logs and check lists: 
\code{\link{check_analysis_dap}()},
\code{\link{check_check_list}()},
\code{\link{check_cleaning_log}()}
}
\concept{functions to check logs and check lists}
