% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_analysis_dap}
\alias{check_analysis_dap}
\title{Check analysis}
\usage{
check_analysis_dap(dap, design, survey, choices)
}
\arguments{
\item{dap}{Some dap}

\item{design}{A \code{srvyr} design}

\item{survey}{The survey sheet}

\item{choices}{The choices sheet}
}
\description{
Check analysis
}
\seealso{
Other functions to check logs and check lists: 
\code{\link{check_analysis}()},
\code{\link{check_check_list}()},
\code{\link{check_cleaning_log}()}
}
\concept{functions to check logs and check lists}
