% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_check_list}
\alias{check_check_list}
\title{Check cleaning log against data}
\usage{
check_check_list(check_list, .tbl)
}
\arguments{
\item{check_list}{A check list (a tibble of a particular format)}

\item{.tbl}{Data to clean}
}
\description{
Check cleaning log against data
}
\details{
Names should be the English version. Please use [impactR::log_names_fr_en] beforehand.
}
\seealso{
Other functions to check logs and check lists: 
\code{\link{check_cleaning_log}()}
}
\concept{functions to check logs and check lists}
