% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_cleaning_log}
\alias{check_cleaning_log}
\title{Check cleaning log}
\usage{
check_cleaning_log(log, .tbl, id_col, other)
}
\arguments{
\item{log}{A cleaning log}

\item{.tbl}{Data to clean}

\item{id_col}{The column id, usually "uuid"}

\item{other}{A character vector of the start of all other column names. E.g., other = "other_"}
}
\description{
Check cleaning log
}
\details{
Names should be the English version. If necessary, please use \link{log_names_fr_en} beforehand.
}
\seealso{
Other functions to check logs and check lists: 
\code{\link{check_check_list}()}
}
\concept{functions to check logs and check lists}
