% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{check_list}
\alias{check_list}
\title{Some table of checks}
\format{
A data frame with a few rows and variables:
\describe{
  \item{id_check}{The check id}
  \item{question_name}{Tne Kobo survey question name}
  ...
}
}
\usage{
check_list
}
\description{
A dataset (in French) containing a list of checks to pass to `make_log()`
}
\keyword{datasets}
