% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cite_cr.R
\name{cite_cr}
\alias{cite_cr}
\title{Derive citation data from the Crossref database via rcrossref}
\usage{
cite_cr(id_list)
}
\arguments{
\item{id_list}{Vector of pmid / doi}
}
\value{
Dataframe
}
\description{
Derive citation data from the Crossref database
}
