% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cite_dim.R
\name{cite_dim}
\alias{cite_dim}
\title{Derive citation data from the Dimentions database (https://app.dimensions.ai/discover/publication)}
\usage{
cite_dim(id_list)
}
\arguments{
\item{id_list}{Vector of pmid / doi}
}
\value{
Dataframe
}
\description{
Derive citation data from the Dimentions database
}
