% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cite_oc.R
\name{cite_oc}
\alias{cite_oc}
\title{Derive citation data from the Open Citations database.}
\usage{
cite_oc(id_list)
}
\arguments{
\item{id_list}{Vector of pmid / doi}
}
\value{
Nested dataframe of (1) All referencing publications (2) Total citations recorded (3) Citations over time
}
\description{
Derive citation data from the Open Citations database (https://opencitations.net/). Note this is too sparce a resource to be of practical use at present.
}
