% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cite_presentation.R
\name{cite_presentation}
\alias{cite_presentation}
\title{Provide a citation for a research presentation}
\usage{
cite_presentation(df, type = "type", level = "level",
  title = "title", con_org = "con_org", con_name = "con_name",
  con_date_start = "con_date_start", con_date_end = "con_date_end",
  con_city = "con_city", con_country = "con_country",
  author = "author",
  cite_format = "author. title. con_org con_name, con_date_range, con_city (con_country).")
}
\arguments{
\item{df}{Dataframe with publications listed rowwise}

\item{type}{String of column name with presentation type (e.g. poster, oral, etc) listed (optional).}

\item{level}{String of column name with presentation level (e.g. national, international, etc) listed (optional).}

\item{title}{String of column name with presentation title listed (mandatory).}

\item{con_org}{String of column name with the name of the organisation where the work was presented (mandatory)}

\item{con_name}{String of column name with the name of the meeting/conference where the work was presented (mandatory)}

\item{con_date_start}{String of column name which represents the start date (optional).}

\item{con_date_end}{String of column name which represents the end date (optional).}

\item{con_city}{String of column name with the city of the meeting/conference where the work was presented (mandatory)}

\item{con_country}{String of column name with the country of the meeting/conference where the work was presented (mandatory)}

\item{author}{String of column name with the authors of the work presented (mandatory).}

\item{cite_format}{Customisable string specifying the exact format for the preferred referencing style. Note exact parameter names are required and are limited to those specified.}
}
\value{
Dataframe with addition of citation variable.
}
\description{
Used to generate a text version of the difference between 2 dates.
}
