% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cite_publication.R
\name{cite_publication}
\alias{cite_publication}
\title{Provide a citation for a research publication}
\usage{
cite_publication(df, author = "author", title = "title",
  journal = "journal", year = "year", volume = "volume",
  issue = "issue", pages = "pages", pmid = "pmid", doi = "doi",
  max_auth = FALSE,
  cite_format = "author. title. journal. year; volume (issue): pages. PMID: pmid. DOI: doi.")
}
\arguments{
\item{df}{Dataframe with publications listed rowwise}

\item{author}{String of column name with publication authors listed (mandatory).}

\item{title}{String of column name with publication title listed (mandatory).}

\item{journal}{String of column name with publication journal name listed (mandatory).}

\item{year}{String of column name with publication year listed (mandatory).}

\item{volume}{String of column name with publication volume listed (mandatory).}

\item{issue}{String of column name with publication issue listed (optional)}

\item{pages}{String of column name with publication pages listed (mandatory).}

\item{pmid}{String of column name with publication pmid listed (optional)}

\item{doi}{String of column name with publication journal name listed (optional).}

\item{max_auth}{Argument to limit number of authors listed to the number specified. All n+1 authors will be replaced by a single "et al.". (default = FALSE e.g. no restriction).}

\item{cite_format}{Customisable string specifying the exact format for the preferred referencing style (default is Vancouver style). Note exact parameter names are required and are limited to those specified.}
}
\value{
Dataframe with addition of citation variable.
}
\description{
Used to generate a text version of the difference between 2 dates.
}
