% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cite_scopus.R
\name{cite_scopus}
\alias{cite_scopus}
\title{Derive citation data from the Crossref database via rcrossref}
\usage{
cite_scopus(id, scopus_api = rscopus::get_api_key())
}
\arguments{
\item{id}{Vector of pmid / doi}

\item{scopus_api}{Either scopus API directly or ignore if preset via rscopus::set_api_key(scopus_api). Instructions to obtain can be found here (https://rdrr.io/cran/rscopus/f/vignettes/api_key.Rmd)}
}
\value{
Dataframe
}
\description{
Derive citation data from the Crossref database
}
