% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_all}
\alias{clean_all}
\title{Clean all}
\usage{
clean_all(.tbl, log, survey, choices, id_col, other)
}
\arguments{
\item{.tbl}{A tibble}

\item{log}{A log, which contains a column "action".}

\item{survey}{A survey sheet from Kobo (with column "type" split)}

\item{choices}{A choices sheet from Kobo}

\item{id_col}{Usually uuid... to count}

\item{other}{A character vector of the start of all other column names. E.g., other = "other_"}
}
\value{
A cleaned tibble
}
\description{
Clean all
}
\details{
Apply all cleaning functions, after checking the log, in the right order, modify character and double variables, recode others and other parents, remove duplicates, remove surveys, recount occurences. It uses default for count_occ_all.
}
