% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workbook.R
\name{clean_workbook_list}
\alias{clean_workbook_list}
\alias{clean_workbook_df}
\title{Export a clean workbook}
\usage{
clean_workbook_list(
  named_list,
  path,
  col_width = 15,
  max_width = 60,
  font_name = "Leelawadee",
  font_size = 10,
  table_style = "TablestyleMedium10"
)

clean_workbook_df(
  df,
  name,
  path,
  col_width = 15,
  max_width = 60,
  font_name = "Leelawadee",
  font_size = 10,
  table_style = "TablestyleMedium10"
)
}
\arguments{
\item{named_list}{A named list of tibbles/dataframes.}

\item{path}{A full full or relative path that includes file name.}

\item{col_width}{Minimum column width. Default to 15.}

\item{max_width}{Maximum column width. Default to 60.}

\item{font_name}{Font name. Default to "Leelawadee".}

\item{font_size}{Font size. Default to 10.}

\item{table_style}{Excel table style. Default to "TablestyleMedium10".}

\item{df}{Dataframe to export}

\item{name}{Character string if the spreadsheet name}
}
\value{
A written xlsx file
}
\description{
The clean_workbook functions exports a list or a dataframe/tibble to nice xlsx file using openxlsx.

* `clean_workbook_list()` takes a named_list as main arg and writes a xlsx file where each list item is a tab.

* `clean_workbook_df()` takes a dataframe as main arg and writes a single-tabbed xlsx file.
}
\section{Functions}{
\itemize{
\item \code{clean_workbook_list}: Export a clean workbook from a named list of dataframes

\item \code{clean_workbook_df}: Export a clean workbook from a single dataframe
}}

