% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comb_name_size.R
\name{comb_name_size}
\alias{comb_name_size}
\title{Additional function for ComplexHeatmap}
\usage{
comb_name_size(comb_mat)
}
\arguments{
\item{comb_mat}{Combination matrix output from ComplexHeatmap::make_comb_mat()}
}
\value{
Tibble of combinations with the name of the sets provided.
}
\description{
Additional function for ComplexHeatmap to faciltate named sets in combinations
}
