% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{count_occ}
\alias{count_occ}
\title{Count occurences of variables}
\usage{
count_occ(.tbl, id_col, col)
}
\arguments{
\item{.tbl}{A tibble}

\item{id_col}{Usually uuid... to count.}

\item{col}{A column of .tbl}
}
\description{
Count occurences of variables
}
\details{
This is solely written to mimic variables names produced by Kobo after use of `janitor::clean_names()`. Example : *variable* is a multiple choice variable and *variable_choice1* is its count variable of choice 1.
}
