% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{cross_freq}
\alias{cross_freq}
\title{Calculate cross-frequencies}
\usage{
cross_freq(.tbl, row, col, group_col = NULL, output = 1)
}
\arguments{
\item{.tbl}{A tibble}

\item{row}{Variable to rows}

\item{col}{Variable to pivot to columns}

\item{group_col}{Variable to group_by. Default to NULL}

\item{output}{Default to 1. What should be the output? 1 for frequencies only, 2 for counts only, 3 for both.}
}
\value{
A cross-table of frequencies
}
\description{
Calculate cross-frequencies
}
