% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{daily_acc}
\alias{daily_acc}
\title{Daily living accelerometer data collected during 3 days}
\format{
A data frame with 25,913,900 rows and 4 variables:
\describe{
\item{timestamp}{The data collection timestamp.}
\item{acc_X, acc_Y, acc_Z}{The acceleration on the X, Y and Z axes,
stored in gravitational acceleration units (\emph{g}).}
}
}
\usage{
daily_acc
}
\description{
A dataset containing the timestamp and acceleration values of three days of
accelerometer data collection using a triaxial accelerometer with a sampling
frequency of 100Hz (GT9X Link, ActiGraph, Pensacola, FL, USA).
}
\keyword{datasets}
