% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dap}
\alias{dap}
\title{Some random dap test}
\format{
A data frame with 5 rows and 13 variables:
\describe{
  \item{question_name}{The question_name or the ratio}
  \item{analysis}{Tne R type of analysis to pass to `make_analysis`}
  ...
}
}
\usage{
dap
}
\description{
A dataset (in French) containing a basic dap, just to test. Very experimental
}
\keyword{datasets}
