% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{Some data produced from a Kobo tool}
\format{
A data frame with 8 rows and 36 variables:
\describe{
  \item{start}{The starting time of the survey}
  \item{end}{The end time of the survey}
  ...
}
}
\usage{
data
}
\description{
A dataset (in French) containing data for consensus, admin, enumerator info, chief of hh, and sanitary
}
\keyword{datasets}
