% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_vec.R
\name{date_to_day_month}
\alias{date_to_day_month}
\title{Date to day and month ("fr" locale)}
\usage{
date_to_day_month(.vec)
}
\arguments{
\item{.vec}{A vector of type POSIXct"}
}
\value{
The same vector with only the day and month in the env locale.
}
\description{
Date to day and month ("fr" locale)
}
