% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{define_region}
\alias{define_region}
\title{Define region of interest}
\usage{
define_region(data, start, end)
}
\arguments{
\item{data}{An \code{impactr_data} object, as obtained with
\link[=read_acc]{read_acc()}.}

\item{start, end}{A character string with the start and end times of the
region of interest in the "HH:MM:SS" format.}
}
\value{
An object of class \code{impactr_data}.
}
\description{
Define the region of interest for data analysis based on the accelerometer
data timestamp.
}
\examples{
data <- read_acc(impactr_example("hip-imu.csv"))
define_region(data, start = "15:08:00", end = "15:45:00")
}
