% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_wear.R
\name{delete_nonwear}
\alias{delete_nonwear}
\title{Delete accelerometer non-wear time}
\usage{
delete_nonwear(data)
}
\arguments{
\item{data}{An \code{impactr_data} object with the \code{wear} column, as
obtained with \link[=mark_nonwear]{mark_nonwear()}.}
}
\value{
The \code{impactr_data} object provided by \code{data} without the
\code{wear} column and with the non-wear time removed.
}
\description{
Deletes all periods marked as non-wear.
}
\details{
This is an internal function, designed to be used inside the wrapper
function \link[=remove_nonwear]{remove_nonwear()} that integrates all
steps of detection, removal and inspection of accelerometer non-wear time.
}
\seealso{
\code{\link[=detect_nonwear]{detect_nonwear()}},
\code{\link[=plot_nonwear]{plot_nonwear()}},
\code{\link[=mark_nonwear]{mark_nonwear()}},
\code{\link[=summarise_nonwear]{summarise_nonwear()}}
}
\keyword{internal}
