% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_df.R
\name{deselect}
\alias{deselect}
\title{Remove cols from tibble}
\usage{
deselect(.tbl, ...)
}
\arguments{
\item{.tbl}{A tibble}

\item{...}{Column names}
}
\value{
A tibble with columns removed
}
\description{
Remove cols from tibble
}
