% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffdate_text.R
\name{diffdate_text}
\alias{diffdate_text}
\title{Provide a text version of the difference between 2 dates.}
\usage{
diffdate_text(df, date_start, date_end)
}
\arguments{
\item{df}{Dataframe.}

\item{date_start}{String of column name in df which represents the start date}

\item{date_end}{String of column name in df which represents the end date.}
}
\value{
Vector of date range as text.
}
\description{
Used to generate a text version of the difference between 2 dates.
}
