% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{encode}
\alias{encode}
\title{Vector encoding (from Latin-1 to UTF-8)}
\usage{
encode(.vec, from = "latin1", to = "UTF-8")
}
\arguments{
\item{.vec}{A character vector (Latin-1 as default)}

\item{from}{The string encoding from. Default is  "latin1"}

\item{to}{The string encoding to. Default is " "UTF-8"}
}
\value{
The same vector but UTF-8
}
\description{
Vector encoding (from Latin-1 to UTF-8)
}
