% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_doi.R
\name{extract_doi}
\alias{extract_doi}
\title{Extract publication data using doi / crossref}
\usage{
extract_doi(doi, get_auth = TRUE, get_altmetric = TRUE, get_impact = TRUE)
}
\arguments{
\item{doi}{Vector of Digital Object Identifiers (DOI)}

\item{get_auth}{Extract authorship data (default = TRUE)}

\item{get_altmetric}{Extract overall altmetric score data (default = TRUE)}

\item{get_impact}{Extract journal impact factor and metrics (default = TRUE)}
}
\value{
Dataframe of essential publication data
}
\description{
Extract publication data using doi / rcrossref
}
