% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_doi_auth.R
\name{extract_doi_auth}
\alias{extract_doi_auth}
\title{Extract publication data on authorship using doi / crossref}
\usage{
extract_doi_auth(doi)
}
\arguments{
\item{doi}{Vector of Digital Object Identifiers (DOI)}
}
\value{
Dataframe authorship data for each DOI.
}
\description{
Extract publication data on authorship using doi / rcrossref
}
