% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_impact_factor.R
\name{extract_impact_factor}
\alias{extract_impact_factor}
\title{Provide a citation for a research publication}
\usage{
extract_impact_factor(df, var_id = "pmid", var_issn = "journal_issn")
}
\arguments{
\item{df}{Dataframe with publications listed rowwise with at least 1 column containing the journal ISSN}

\item{var_issn}{String of column name with the journal International Standard Serial Number (ISSN) number listed (default: journal_issn)}
}
\value{
Dataframe of journal impact metrics
}
\description{
Used to extract journal impact metrics for publications.
}
