% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_scholar_journal.R
\name{extract_scholar_journal}
\alias{extract_scholar_journal}
\title{Extract additional jounral metrics from Google Scholar}
\usage{
extract_scholar_journal(list_journals)
}
\arguments{
\item{list_journals}{Vector of journal names (note: Google Scholar requires specific formats to match to journal records)}
}
\value{
Nested dataframe of: (1). "metrics" - journal metrics derived from Google Scholar; (2). "unmatch" - list of journals unable to be matched within Google Scholar.
}
\description{
Extract additional jounral metrics from Google Scholar
}
