% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators_foodsec.R
\name{fclcp}
\alias{fclcp}
\title{FEWS NET Food Consumption-Livelihood Coping Matrix}
\usage{
fclcp(.tbl, fcp, lcs_cat)
}
\arguments{
\item{.tbl}{Data}

\item{fcp}{The food consumption phase column produced by \code{impactR::fcp()}}

\item{lcs_cat}{The Livelihood Coping Strategies categories produced by \code{impactR::lcs()}}
}
\value{
Two new columns with the FEWS NET Food Consumption-Livelihood Coping matrix phases: numbers (fclp) and categories (fclp_cat).
}
\description{
FEWS NET Food Consumption-Livelihood Coping Matrix
}
