% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators_foodsec.R
\name{fcp}
\alias{fcp}
\title{FC phase - Food Consumption phase according to the FCM}
\usage{
fcp(.tbl, fcm_cell)
}
\arguments{
\item{.tbl}{Data}

\item{fcm_cell}{Column produced by `impactR::fcm()`}
}
\value{
Two columns: the food consumption phase number (fcp) and the category (fcp_cat)
}
\description{
FC phase - Food Consumption phase according to the FCM
}
