% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_df.R
\name{filter2_equal}
\alias{filter2_equal}
\title{Extract a subset with one equal_to_value filter or two}
\usage{
filter2_equal(df, col, value, ..., col2 = NULL, value2 = NULL)
}
\arguments{
\item{df}{A data frame.}

\item{col}{A column.}

\item{value}{A value.}

\item{...}{cols to keep, if none, return all columns}

\item{col2}{A second column. Default to NULL.}

\item{value2}{A second value. Default to NULL.}
}
\value{
A filtered data frame
}
\description{
Extract a subset with one equal_to_value filter or two
}
