% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_df.R
\name{filter2_nequal}
\alias{filter2_nequal}
\title{Extract a subset with one unequal_to_value filter or two}
\usage{
filter2_nequal(df, col, value, col2 = NULL, value2 = NULL, ...)
}
\arguments{
\item{df}{A data frame.}

\item{col}{A column.}

\item{value}{A value.}

\item{col2}{A second column. Default to NULL.}

\item{value2}{A second value. Default to NULL.}

\item{...}{cols to keep, if none, return all columns}
}
\value{
A filtered data frame
}
\description{
Extract a subset with one unequal_to_value filter or two
}
