% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal.R
\name{filter_signal}
\alias{filter_signal}
\title{Filter a signal}
\usage{
filter_signal(signal, n, wn, type)
}
\arguments{
\item{signal}{A numeric vector.}

\item{n}{The order of the filter.}

\item{wn}{The critical frequency or frequencies.}

\item{type}{The type of the filter.}
}
\value{
A numeric vector with the filtered signal.
}
\description{
Apply a butterworth digital filter to a signal.
}
