% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_pubmed_xml.R
\name{format_pubmed_xml}
\alias{format_pubmed_xml}
\title{Extract all pubmed data for list of PMID}
\usage{
format_pubmed_xml(
  pmid,
  var_id = TRUE,
  var_journal = TRUE,
  var_author = TRUE,
  var_collaborator = TRUE,
  var_metadata = TRUE,
  var_history = TRUE,
  var_registry = F,
  var_abstract = F,
  n_chunk = 200
)
}
\arguments{
\item{var_id}{Extract article identification numbers (PMID, DOI, and PMC)}

\item{var_journal}{Extract journal information (journal name and abbreviation, journal identification numbers, journal issue/volume/pages)}

\item{var_author}{Extract author information  (names with and without affiliations, overall number)}

\item{var_collaborator}{Extract collaborator information (names with and without affiliations, overall number)}

\item{var_metadata}{Extract article metadata (publication status, type of article)}

\item{var_history}{Extract article history (dates of publication, entry onto pubmed, etc)}

\item{var_registry}{Extract article registry information (registry number and location)}

\item{var_abstract}{Extract article abstract}

\item{n_chunk}{Maximum number within each individual API request to pubmed (suggested <200).}

\item{xml_pubmed}{Vector of PMID that details are desired from.}
}
\value{
Tibble of data extracted from pubmed XML
}
\description{
Extract all pubmed data for list of PMID
}
