% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_kobo.R
\name{get_choices}
\alias{get_choices}
\title{Get choices from survey name (and paste them if you want!)}
\usage{
get_choices(survey, choices, col, conc = T, label = F)
}
\arguments{
\item{survey}{A survey sheet from Kobo (already split with columns list_name and name present)}

\item{choices}{A choices sheet from Kobo (with column list_name, label and name)}

\item{col}{A quoted column name}

\item{conc}{Should choices be concatenated to column name? Default to TRUE. Can only be used together swith label = F.}

\item{label}{Should the labels be returned?}
}
\value{
A character vector of choices or pasted to \code{col} choices with sep "_"
}
\description{
Get choices from survey name (and paste them if you want!)
}
