% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_other_parent}
\alias{get_other_parent}
\title{Get all select multiple and parent other questions}
\usage{
get_other_parent(.tbl, other)
}
\arguments{
\item{.tbl}{A tibble of data}

\item{other}{A character vector of the start of all other column names. E.g., other = "other_"}
}
\value{
A character vector of other questions
}
\description{
Get all select multiple and parent other questions
}
\details{
This function has a tibble and the associated survey sheet as inputs, as well as the beginning of the "other" character string. It returns all columns that exist in the tibble and are either a multiple choice or a parent other question. This allows then to compute these columns with `impactR::count_occ` after parent question may have been recoded.
}
