% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_kobo.R
\name{get_select_multiple_and_other_parent}
\alias{get_select_multiple_and_other_parent}
\title{Get all select multiple and parent other questions}
\usage{
get_select_multiple_and_other_parent(.tbl, survey, other)
}
\arguments{
\item{.tbl}{A tibble of data}

\item{survey}{The survey sheet from Kobo (with column "type" split)}

\item{other}{A character vector of the start of all other column names. E.g., other = "other_"}
}
\value{
A character vector of select_multiple and other questions
}
\description{
Get all select multiple and parent other questions
}
\details{
This function has a tibble and the associated survey sheet as inputs, as well as the beginning of the "other" character string. It returns all columns that exist in the tibble and are either a multiple choice or a parent other question. This allows then to compute these columns with \code{impactR::count_occ} after parent question may have been recoded.
}
