% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_kobo.R
\name{get_select_one}
\alias{get_select_one}
\title{Get all select one / simple choices from survey sheet}
\usage{
get_select_one(survey)
}
\arguments{
\item{survey}{The survey sheet from Kobo (with column "type" split)}
}
\value{
A character vector of select_one questions
}
\description{
Get all select one / simple choices from survey sheet
}
\details{
survey should have a split type column with types of variables such as "select_one", "select_multiple", etc.
}
