% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{get_survey_label}
\alias{get_survey_label}
\title{Get survey label from question name}
\usage{
get_survey_label(.tbl, survey, col_to_join)
}
\arguments{
\item{.tbl}{A tibble with at least a column made of Kobo survey names}

\item{survey}{The survey}

\item{col_to_join}{Column character string to join}
}
\value{
A tibble with question_label
}
\description{
Get survey label from question name
}
\details{
It is mainly used in the make_log* functions. E.g., after `bind_others`
}
