% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators_foodsec.R
\name{hhs}
\alias{hhs}
\title{HHS - Household Hunger Scale}
\usage{
hhs(
  .tbl,
  hhs_lev1_nofoodhh,
  hhs_lev2_nofoodhh,
  hhs_lev1_sleephungry,
  hhs_lev2_sleephungry,
  hhs_lev1_alldaynight,
  hhs_lev2_alldaynight,
  level1_codes,
  level2_codes
)
}
\arguments{
\item{.tbl}{Data}

\item{hhs_lev1_nofoodhh}{Component column: No food of any kind in the house}

\item{hhs_lev2_nofoodhh}{Follow-up frequency column}

\item{hhs_lev1_sleephungry}{Component column: Go to sleep hungry because there was not enough food}

\item{hhs_lev2_sleephungry}{Follow-up frequency column}

\item{hhs_lev1_alldaynight}{Component column: Go a whole day and night without eating}

\item{hhs_lev2_alldaynight}{Follow-up frequency column}

\item{level1_codes}{Character vector of at least "Yes" and "No" codes (in this order), e.g. c("yes", "no")}

\item{level2_codes}{Character vector of at least frequencies codes, in the following order: "Rarely", "Sometimes", "Often", e.g. c("rarely", "sometimes", "often")}
}
\value{
Five new columns: each component recoded hhs_comp1, hhs_comp2, hhs_comp3, the overall score (hhs_score) and categories (hhs_cat).
}
\description{
HHS - Household Hunger Scale
}
