% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{if_not_in_stop}
\alias{if_not_in_stop}
\title{Stop statement "If not in colnames" with colnames}
\usage{
if_not_in_stop(.tbl, cols, df, arg = NULL)
}
\arguments{
\item{.tbl}{A tibble}

\item{cols}{A vector of column names (quoted)}

\item{df}{Provide the tibble name as a character string}

\item{arg}{Default to NULL.}
}
\value{
A stop statement
}
\description{
Stop statement "If not in colnames" with colnames
}
