% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{if_vec_not_in_stop}
\alias{if_vec_not_in_stop}
\title{Stop statement "If not in vector"}
\usage{
if_vec_not_in_stop(vec, cols, vec_name, arg = NULL)
}
\arguments{
\item{vec}{A vector of character strings}

\item{cols}{A vector of column names (quoted)}

\item{vec_name}{Provide the vector name as a character string}

\item{arg}{Default to NULL.}
}
\value{
A stop statement
}
\description{
Stop statement "If not in vector"
}
