% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_almetric.R
\name{impact_almetric}
\alias{impact_almetric}
\title{Extract almetric data on social media engagment}
\usage{
impact_almetric(list_pmid)
}
\arguments{
\item{list_pmid}{Vector of unique PubMed identifier numbers (PMID)}
}
\value{
Dataframe of data on social media engagment from the almetric repository
}
\description{
Extract impact data from social media engagment
}
