% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_altmetric.R
\name{impact_altmetric}
\alias{impact_altmetric}
\title{Extract altmetric data on social media engagment}
\usage{
impact_altmetric(list_pmid)
}
\arguments{
\item{list_pmid}{Vector of unique PubMed identifier numbers (PMID)}
}
\value{
Nested dataframe (1) df_output; Original dataset with appended altmetric data (2) temporal; Long format data on temporal changes (3) rank; Long format data altmetric ranking (4) source; Long format data sources used to derive altmetric.
}
\description{
Extract impact data from social media engagment
}
