% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_auth.R
\name{impact_auth}
\alias{impact_auth}
\title{Provide a citation for a research publication}
\usage{
impact_auth(
  data,
  author_list = "author_list",
  pub_group = "pmid",
  max_inital = 1,
  upset = FALSE,
  metric = FALSE
)
}
\arguments{
\item{author_list}{A vector of the collapsed authorship lists of each publication, listed per row (should be an output from paste0(x, collapse = ", ")).}

\item{pub_group}{A vector of the research outputs (the paper or project authored) which the authors are being compared across (default = pmid)}

\item{max_inital}{= Number of initials to match authors on (default = 1).}

\item{upset}{Compare intersections of authorship across each pub_group, and produces data format suitable for producing UpSet plots (default = FALSE).}

\item{metric}{Provide author engagement metrics across each pub_group (default = FALSE)}

\item{df}{Dataframe with 2 mandatory columns (1) pub_group (the paper or project authored) (2) auth_list = A string of all authors for the pub_group.}
}
\value{
Nested dataframe of: (1). "auth_out" - All unique authors; (2). data_upset = intersection data (if upset = TRUE). (3). metric = author engagement metrics (if metric = TRUE).
}
\description{
Used to generate a text version of the difference between 2 dates.
}
