% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_network.R
\name{impact_auth_network}
\alias{impact_auth_network}
\title{Create a dataframe for network analysis of co-authorship}
\usage{
impact_auth_network(
  df,
  author = "author_list",
  id = "pubmed",
  auth_interest = "",
  initial_right = TRUE,
  initial_num = 1,
  edge_min = 1
)
}
\arguments{
\item{df}{Dataframe with 2 mandatory columns (1) "id": A vector of unique paper IDs (e.g. DOI / PMID) (2) "author": A vector of strings of all authors (format must be last name + initials)}

\item{author}{Name of the "author" variable in the dataframe (default="author_list")}

\item{id}{Name of the "id" variable in the dataframe (default="pubmed")}

\item{auth_interest}{= List of authors of interest (will exclude all vertices *not* involving these authors)}

\item{initial_right}{= Are the initials to the right of the last name? (default = TRUE)}

\item{edge_min}{= The minimum number of edges (weight) desired (default = 1)}

\item{initial_n}{= Number of initials to match authors on (default = 1).}
}
\value{
Nested dataframe of: (1) Nodes (2) Vertices for network analysis
}
\description{
Create a dataframe for network analysis of co-authorship
}
