% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESCRIPTION.R
\docType{package}
\name{impactR}
\alias{impactR}
\title{Bring some survey analysis functions and some utils.}
\description{
Mainly a reminder of totally-perfectible functions used and made on the go
for the Burkina Faso mission in 2021. Some functions written as utils (to never
write them again) and to monitor, clean and analyze (still to be written) Kobo
survey data, mainly using the \code{srvyr} package and the tidyverse overall.
It requires R 4.1+ (mostly for the native pipe \verb{|>}).
}
\author{
Guillaume Noblet \email{gnoblet@zaclys.net}
}
