% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impactr_example.R
\name{impactr_example}
\alias{impactr_example}
\title{Get path to example data}
\usage{
impactr_example(file = NULL)
}
\arguments{
\item{file}{A character string with the file name. If \code{NULL}, the
example files will be listed.}
}
\value{
If \code{file = NULL}, it returns the file names of the example
data files, else it returns the path to the example data.
}
\description{
\code{impactr} comes with some example ActiGraph accelerometer raw data
files in its \code{inst/extdata} directory. This function make them easy to
access.
}
\examples{
impactr_example()
impactr_example("hip-raw.csv")
}
