% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer.R
\name{infer_impact}
\alias{infer_impact}
\title{Infer Impact Using Causal Models}
\usage{
infer_impact(data, logic_model, method = "causalforest")
}
\arguments{
\item{data}{A data.frame with outcome, treatment, and covariates}

\item{logic_model}{A logic model list}

\item{method}{The method to use ('causalforest')}
}
\value{
GRF prediction object
}
\description{
Infer Impact Using Causal Models
}
