% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{install_accdata}
\alias{install_accdata}
\title{Install accdata package}
\usage{
install_accdata()
}
\description{
A helper function to install the \code{accdata} package from a
\href{https://lveras.com/drat/}{\code{drat}} repository. The \code{accdata}
package contains datasets that can be used to test the functionalities
from \code{impactr}. Note that \code{accdata} is a large package
(approximately 80 MB) and could take a while to download and install.
}
