% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_exports.R
\name{install_scipy}
\alias{install_scipy}
\title{Install SciPy}
\usage{
install_scipy(method = "auto", conda = "auto")
}
\arguments{
\item{method}{Installation method. By default, "auto" automatically finds a
method that will work in the local environment.}

\item{conda}{The path to a conda executable. Use "auto" to allow reticulate
to automatically find an appropriate conda binary.}
}
\description{
Install the SciPy Python package into a virtual environment.
}
\seealso{
\code{\link[reticulate:py_install]{py_install()}}
}
