% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impactr_data.R
\name{is_impactr}
\alias{is_impactr}
\alias{is_impactr_data}
\alias{is_impactr_peaks}
\title{Test if the object is from the impactr package}
\usage{
is_impactr_data(x)

is_impactr_peaks(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if the object inherits the class being evaluated.
}
\description{
Test if the object is from the impactr package
}
