% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_kobo.R
\name{label_all_select_multiple}
\alias{label_all_select_multiple}
\title{Label all select_multiple questions}
\usage{
label_all_select_multiple(data, survey, choices, id_col)
}
\arguments{
\item{data}{Some kobo data}

\item{survey}{Some survey sheet, with a split 'type' column, e.g. with \code{split_survey(type)}. It must have columns 'list_name' and 'name'}

\item{choices}{The corresponding choices sheet; it must have columns name and label}

\item{id_col}{The id column; usually uuid}
}
\value{
Select-multiple labelled dataframe
}
\description{
Label all select_multiple questions
}
