% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{label_select_one}
\alias{label_select_one}
\title{Label select_one question}
\usage{
label_select_one(data, survey, choices, id_col, col, return_df = NULL)
}
\arguments{
\item{data}{Some kobo data}

\item{survey}{Some survey sheet, with a split 'type' column, e.g. with `split_survey(type)`. It must have columns 'list_name' and 'name'}

\item{choices}{The corresponding choices sheet; it must have columns name and label}

\item{id_col}{The id column; usually uuid}

\item{col}{The select_one column to labelize}

\item{return_df}{Default returns the updated dataframe; "id_col" returns a dataframe with col and id_col; and, "vector" returns}
}
\value{
A labelled dataframe, sub-dataframe or vector
}
\description{
Label select_one question
}
