% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators_foodsec.R
\name{lcsi}
\alias{lcsi}
\title{LCSI - Livelihood Coping Strategy Index}
\usage{
lcsi(
  .tbl,
  lcs_stress_1,
  lcs_stress_2,
  lcs_stress_3,
  lcs_stress_4,
  lcs_crisis_1,
  lcs_crisis_2,
  lcs_crisis_3,
  lcs_emergency_1,
  lcs_emergency_2,
  lcs_emergency_3,
  level_codes
)
}
\arguments{
\item{.tbl}{Data}

\item{lcs_stress_1}{Component column: Stress strategy 1}

\item{lcs_stress_2}{Component column: Stress strategy 2}

\item{lcs_stress_3}{Component column: Stress strategy 3}

\item{lcs_stress_4}{Component column: Stress strategy 4}

\item{lcs_crisis_1}{Component column: Crisis strategy 1}

\item{lcs_crisis_2}{Component column: Crisis strategy 2}

\item{lcs_crisis_3}{Component column: Crisis strategy 3}

\item{lcs_emergency_1}{Component column: Emergency strategy 1}

\item{lcs_emergency_2}{Component column: Emergency strategy 2}

\item{lcs_emergency_3}{Component column: Emergency strategy 3}

\item{level_codes}{Character vector of at responses codes, in the following order: ""Yes", "No, exhausted", "No, no need", "No, not applicable", e.g. c("yes", "exhausted", "no_need", "not_applicable")}
}
\value{
Fourteen new columns: each strategy recoded (lcs_stress_*, lcs_crisis_*, lcs_emergency_*), a dummy for each category (lcs_stress, lcs_crisis, lcs_emergency), and the category (lcs_cat).
}
\description{
LCSI - Livelihood Coping Strategy Index
}
