% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_df.R
\name{left_joints}
\alias{left_joints}
\title{Left join a few tibble from a list}
\usage{
left_joints(list, ...)
}
\arguments{
\item{list}{A list of tibbles}

\item{...}{Grouping columns}
}
\value{
A left-joined tibble
}
\description{
Left join a few tibble from a list
}
