% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{log_names_en_fr}
\alias{log_names_en_fr}
\title{Rename log columns from English to French}
\usage{
log_names_en_fr(.tbl)
}
\arguments{
\item{.tbl}{A tibble cleaning log (French names)}
}
\description{
Rename log columns from English to French
}
