% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{log_names_fr_en}
\alias{log_names_fr_en}
\title{Rename log columns from French to English}
\usage{
log_names_fr_en(.tbl)
}
\arguments{
\item{.tbl}{A tibble cleaning log (French names)}
}
\description{
Rename log columns from French to English
}
