% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{make_all_logs}
\alias{make_all_logs}
\title{Make all logs}
\usage{
make_all_logs(.tbl, survey, check_list, other, id_col, ...)
}
\arguments{
\item{.tbl}{A tibble of data}

\item{survey}{The survey sheet of the Kobo tool}

\item{check_list}{A tibble of logical test to check for}

\item{other}{A character vector of the start of all other column names. E.g., other = "other_"}

\item{id_col}{Survey id, usually "uuid"}

\item{...}{Columns to keep in the log, e.g. today, i_enum_id}
}
\value{
A full log as a tibble
}
\description{
Make all logs
}
\seealso{
\link{make_log} for a log based on a logical test

\code{\link[=make_log_from_check_list]{make_log_from_check_list()}} for a log based on many logical tests

\code{\link[=make_log_outlier]{make_log_outlier()}} for a log based on outliers

\code{\link[=make_log_other]{make_log_other()}} for a log of others

Other functions to make logs: 
\code{\link{make_log_from_check_list}()},
\code{\link{make_log_other}()},
\code{\link{make_log_outlier}()},
\code{\link{make_log}()}
}
\concept{functions to make logs}
