% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{make_analysis}
\alias{make_analysis}
\title{Make analysis}
\usage{
make_analysis(
  design,
  survey,
  choices,
  col,
  analysis,
  group = NULL,
  level = 0.9,
  na_rm = T,
  vartype = "ci"
)
}
\arguments{
\item{design}{A design object}

\item{survey}{The survey sheet from Kobo that contains at least column 'list_name' (split from 'type') and 'name'}

\item{choices}{The choices sheet from Kobo contains at least column 'list_name' (split from 'type') and 'name'}

\item{col}{Column to make analysis from}

\item{analysis}{One of "median", "mean", "prop_simple", "prop_multiple", "prop_multiple_overall", "ratio"}

\item{group}{Variable(s) to group by}

\item{level}{Confidence level to pass to `svy_*` functions}

\item{na_rm}{Should NAs be removed prior to calculation ?}

\item{vartype}{Parameter from `srvyr` functions. Default to "ci"}
}
\value{
A summarized analysis
}
\description{
Make analysis
}
\details{
The design object must have had columns with underscores as choices separators, for instance they could have been imported either with `impactR::write_xlsx` or with `janitor::clean_names`

#' #' @description
`r lifecycle::badge("experimental")`

This function still is experimental.
}
