% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{make_log}
\alias{make_log}
\title{Make log}
\usage{
make_log(
  .tbl,
  survey,
  id_col = "uuid",
  cols_to_keep,
  id_check = "empty",
  question_name,
  why = "empty",
  logical_test,
  new_value,
  action,
  type
)
}
\arguments{
\item{.tbl}{A tibble of data}

\item{survey}{The survey sheet from the Kobo tool}

\item{id_col}{Survey id, usually "uuid"}

\item{cols_to_keep}{Columns to keep in the log, e.g. c("today", "i_enum_id")}

\item{id_check}{A string to identify the check}

\item{question_name}{A column of .tbl}

\item{why}{A long character string}

\item{logical_test}{A character string made of a logical expression}

\item{new_value}{Should we suggest a new value?}

\item{action}{A character with "check" as default}

\item{type}{The type of the question "double" or "character"}
}
\value{
A tibble with a log for one question and one logical test
}
\description{
Make log
}
\seealso{
[make_log_from_check_list()] for a log based on many logical tests

[make_log_outlier()] for a log based on outliers

[make_log_other()] for a log of others

[make_all_logs()] for all logs: outliers, others, logical tests

Other functions to make logs: 
\code{\link{make_all_logs}()},
\code{\link{make_log_from_check_list}()},
\code{\link{make_log_other}()},
\code{\link{make_log_outlier}()}
}
\concept{functions to make logs}
