% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{make_log_from_check_list}
\alias{make_log_from_check_list}
\title{Apply \code{make_log} to a tibble of check_list (one logical test per row)}
\usage{
make_log_from_check_list(.tbl, survey, check_list, id_col, ...)
}
\arguments{
\item{.tbl}{A tibble of data}

\item{survey}{The survey sheet of the Kobo tool}

\item{check_list}{A tibble of logical test to check for}

\item{id_col}{Survey id, usually uuid}

\item{...}{Columns to keep in the log, e.g. today, i_enum_id}
}
\description{
Apply \code{make_log} to a tibble of check_list (one logical test per row)
}
\seealso{
\link{make_log} for a log based on a logical test

\code{\link[=make_log_outlier]{make_log_outlier()}} for a log based on outliers

\code{\link[=make_log_other]{make_log_other()}} for a log of others

\code{\link[=make_all_logs]{make_all_logs()}} for all logs: outliers, others, logical tests

Other functions to make logs: 
\code{\link{make_all_logs}()},
\code{\link{make_log_other}()},
\code{\link{make_log_outlier}()},
\code{\link{make_log}()}
}
\concept{functions to make logs}
