% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{make_log_other}
\alias{make_log_other}
\title{Make "other questions" log}
\usage{
make_log_other(.tbl, survey, other, id_col, ...)
}
\arguments{
\item{.tbl}{A tibble of data}

\item{survey}{The survey sheet from Kobo}

\item{other}{A character vector of the start of all other column names. E.g., other = "other_".}

\item{id_col}{Survey id, usually uuid}

\item{...}{Columns to keep in the log, e.g. uuid, enum_id}
}
\value{
A log with all "other_cols".
}
\description{
Make "other questions" log
}
\details{
It assumes that a parent question and its children "other" are coded as follows
within the kobo tool : e.g., parent_question, other_parent_question.
}
\seealso{
\link{make_log} for a log based on a logical test

\code{\link[=make_log_from_check_list]{make_log_from_check_list()}} for a log based on many logical tests

\code{\link[=make_log_outlier]{make_log_outlier()}} for a log based on outliers

\code{\link[=make_all_logs]{make_all_logs()}} for all logs: outliers, others, logical tests

Other functions to make logs: 
\code{\link{make_all_logs}()},
\code{\link{make_log_from_check_list}()},
\code{\link{make_log_outlier}()},
\code{\link{make_log}()}
}
\concept{functions to make logs}
