% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{make_log_outlier}
\alias{make_log_outlier}
\title{Make outlier log}
\usage{
make_log_outlier(.tbl, survey, id_col, ...)
}
\arguments{
\item{.tbl}{A tibble of data}

\item{survey}{The survey sheet from Kobo}

\item{id_col}{Survey id, usually uuid}

\item{...}{Columns to keep in the log, e.g, today, enum_id}
}
\value{
A log with outliers
}
\description{
Make outlier log
}
\details{
It uses both interquartile range (1.5 rule) and standard deviation from mean (3 times rule) to look for outliers
}
\seealso{
[make_log] for a log based on a logical test

[make_log_from_check_list()] for a log based on many logical tests

[make_log_other()] for a log of others

[make_all_logs()] for all logs: outliers, others, logical tests

Other functions to make logs: 
\code{\link{make_all_logs}()},
\code{\link{make_log_from_check_list}()},
\code{\link{make_log_other}()},
\code{\link{make_log}()}
}
\concept{functions to make logs}
