% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_df.R
\name{na_count}
\alias{na_count}
\title{Count the number of occurrences of NAs over a data frame row-wise}
\usage{
na_count(df, new_colname = "count_na")
}
\arguments{
\item{df}{A dataframe}

\item{new_colname}{The newly-mutated colname. Default to "count".}
}
\value{
A mutated dataframe
}
\description{
Count the number of occurrences of NAs over a data frame row-wise
}
