% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_df.R
\name{named_group_split}
\alias{named_group_split}
\title{Named group split}
\usage{
named_group_split(.tbl, group)
}
\arguments{
\item{.tbl}{A tibble of data}

\item{group}{Column to split group by}
}
\value{
A split and named list of tibbles
}
\description{
Named group split
}
