% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impactr_data.R
\name{new_impactr_peaks}
\alias{new_impactr_peaks}
\title{Constructor for impactr_peaks object}
\usage{
new_impactr_peaks(
  x,
  start_date_time,
  samp_freq,
  acc_placement,
  subj_body_mass,
  filter_type,
  peaks_idx,
  curve_start,
  acc_signal
)
}
\arguments{
\item{x}{A data frame.}

\item{start_date_time}{A scalar of class \code{POSIXct}.}

\item{samp_freq}{A numerical scalar}

\item{acc_placement}{A character scalar}

\item{subj_body_mass}{A numerical scalar}

\item{filter_type}{A character scalar}

\item{peaks_idx}{A numerical vector}

\item{curve_start}{A numerical vector}

\item{acc_signal:}{A numeric vector}
}
\value{
An object of the \code{impactr_peaks} class.
}
\description{
Constructor for impactr_peaks object
}
\keyword{internal}
