% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{numeric_cols}
\alias{numeric_cols}
\title{Get numeric colum names}
\usage{
numeric_cols(.tbl, survey = NULL)
}
\arguments{
\item{.tbl}{A tibble}

\item{survey}{Default to NULL. The survey sheet from Kobo.}
}
\value{
A characater vector of numeric columns names
}
\description{
Get numeric colum names
}
\details{
If the survey sheet is provided, it returns numeric columns from \code{.tbl} that are coded as decimal or numeric in the \code{type} column of \code{survey}. Otherwise, it returns all numeric columns from \code{.tbl}.
}
