% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{other_cols}
\alias{other_cols}
\title{Get all other values}
\usage{
other_cols(.tbl, other, id_col)
}
\arguments{
\item{.tbl}{A tibble of data}

\item{other}{A character vector of the start of all other column names. E.g., other = "other_"}

\item{id_col}{Survey id, usually "uuid}
}
\value{
A tibble with all "other" answers
}
\description{
Get all other values
}
