% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{other_parent_cols}
\alias{other_parent_cols}
\title{Get all other values and parent values}
\usage{
other_parent_cols(.tbl, other_cols, other, id_col)
}
\arguments{
\item{.tbl}{A tibble of data}

\item{other_cols}{A tibble produced by `other_cols`}

\item{other}{A character vector of the start of all other column names. E.g., other = "other_"}

\item{id_col}{Unquoted survey id, usually uuid. Need some fix to be used also with quoted names, with the use of `dplyr::group_split`}
}
\value{
A tibble with all "other" values and their parent values
}
\description{
Get all other values and parent values
}
