% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{outliers_sd}
\alias{outliers_sd}
\title{Get outliers outside standard deviations from the mean}
\usage{
outliers_sd(.tbl, col, times = 3, id_col)
}
\arguments{
\item{.tbl}{A tibble}

\item{col}{A numeric columns}

\item{times}{How many standard deviations from the mean? Default to 3}

\item{id_col}{Usually uuid}
}
\value{
A tibble of outliers
}
\description{
Get outliers outside standard deviations from the mean
}
