% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{pull_uuid}
\alias{pull_uuid}
\title{Pull uuid from a logical test}
\usage{
pull_uuid(.tbl, logical_test = "uuid", id_col = "uuid")
}
\arguments{
\item{.tbl}{A tibble}

\item{logical_test}{A logical test as a character expression}

\item{id_col}{Survey id, usually "uuid".}
}
\value{
A vector of ids
}
\description{
Pull uuid from a logical test
}
