% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators_foodsec.R
\name{rcsi}
\alias{rcsi}
\title{rCSI - reduced Coping Strategy Index}
\usage{
rcsi(
  .tbl,
  rcsi_lesspreferred,
  rcsi_borrowfood,
  rcsi_limitportion,
  rcsi_restrict,
  rcsi_reducemeals
)
}
\arguments{
\item{.tbl}{Data}

\item{rcsi_lesspreferred}{Component column: Rely on less preferred and less expensive food}

\item{rcsi_borrowfood}{Component column: Borrow food or rely on help from friends or relatives}

\item{rcsi_limitportion}{Component column: Limit portion size at mealtime}

\item{rcsi_restrict}{Component column: Restrict consumption by adults in order for small children to eat}

\item{rcsi_reducemeals}{Component column: Reduce the number of meals eaten in a day}
}
\value{
Seven new columns: each component weight (rcsi_w_*), rCSI score (rcsi_score) and categories (rcsi_cat).
}
\description{
rCSI - reduced Coping Strategy Index
}
\details{
All component columns must of type 'double'.
}
