% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{reach_colors}
\alias{reach_colors}
\title{Some colors for REACH plots and maps}
\usage{
reach_colors(colors = NULL)
}
\arguments{
\item{colors}{A character string among impactR::reach_colors()}
}
\value{
A named list of individual colors or palettes
}
\description{
Some colors for REACH plots and maps
}
