% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{reach_cols}
\alias{reach_cols}
\title{Function to extract reach colors as hex codes}
\usage{
reach_cols(...)
}
\arguments{
\item{...}{Character names of reach colors. If NULL? returns all colors.}
}
\value{
An hex code
}
\description{
Function to extract reach colors as hex codes
}
\details{
This function needs to be modified to add colors
}
