% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{reach_hist_bar}
\alias{reach_hist_bar}
\title{Simple bar graph}
\usage{
reach_hist_bar(
  .tbl,
  x,
  y = NULL,
  f = NULL,
  c = NULL,
  facet = NULL,
  pal = NULL,
  reverse = F,
  discrete = T,
  alpha = 0.7,
  width = 0.3,
  scale_percent = F,
  position = "fill",
  dodge_width = 0.5,
  fill_name = NULL,
  fill_levels = NULL,
  fill_labels = NULL,
  theme = NULL,
  title = NULL,
  xlab = NULL,
  ylab = NULL
)
}
\arguments{
\item{.tbl}{A tibble}

\item{x}{The x variable}

\item{y}{The y variable}

\item{f}{The fill variable}

\item{c}{The color variable}

\item{facet}{The facet variable}

\item{pal}{The palette name from `impactR::reach_pal()`}

\item{reverse}{Should color be reversed? Default to FALSE}

\item{discrete}{Should color pal be used as a discrete scale? Default to TRUE}

\item{alpha}{Color's transparency}

\item{width}{Bar's width}

\item{scale_percent}{Should axis labels be scaled to nice percent? Default to FALSE}

\item{position}{Either 'fill' or 'dodge'}

\item{dodge_width}{If position is 'dodge', this is the dodge width}

\item{fill_name}{Legend title}

\item{fill_levels}{Levels to order fill labels}

\item{fill_labels}{Fill labels}

\item{theme}{Theme name. Default is NULL to reach theme. Other values are "simple", "void", "bw", "minimal"}

\item{title}{Title}

\item{xlab}{X-axis label}

\item{ylab}{Y-axis label}
}
\description{
Simple bar graph
}
