% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{reach_pal}
\alias{reach_pal}
\title{Return function to interpolate a reach color palette}
\usage{
reach_pal(palette = "main", reverse = FALSE, color_ramp_palette = F, ...)
}
\arguments{
\item{palette}{Character name of palette in drsimonj_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{color_ramp_palette}{Should the output be a `grDevices::colorRampPalette` function or a vector of hex codes? Default to the former with TRUE.}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\value{
A color palette
}
\description{
Return function to interpolate a reach color palette
}
