% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{recode_other_from_log}
\alias{recode_other_from_log}
\title{Recode "other_" from log}
\usage{
recode_other_from_log(.tbl, log, id_col, other)
}
\arguments{
\item{.tbl}{A tibble}

\item{log}{A cleaning log}

\item{id_col}{The column id, usually "uuid"}

\item{other}{A character vector of the start of all other column names. E.g., other = "other_"}
}
\value{
A tibble with "other_" recoded
}
\description{
Recode "other_" from log
}
