% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_df.R
\name{recode_values}
\alias{recode_values}
\title{Replace values of some tibble's columns to one value}
\usage{
recode_values(.tbl, values, to_value, ...)
}
\arguments{
\item{.tbl}{A tibble}

\item{values}{A vector of values}

\item{to_value}{A value}

\item{...}{Column names}
}
\value{
A tibble with recoded values to one value
}
\description{
Replace values of some tibble's columns to one value
}
\details{
If the column type is a character and the replacement a numeric, then the numeric is coerced to a character. If the column type is a numeric and the replacement is a character, then the column is coerced to character. NAs will remains NAs of the right type.
}
