% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{remove_duplicate}
\alias{remove_duplicate}
\title{Remove duplicates}
\usage{
remove_duplicate(.tbl, log, id_col)
}
\arguments{
\item{.tbl}{A tibble}

\item{log}{A deletion log, which contains a column "action".}

\item{id_col}{The column id, usually "uuid"}
}
\value{
Removal - duplicates
}
\description{
Removes duplicated surveys from .tbl where column "action" in the log is set to "duplicate".
}
