% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{remove_from_log}
\alias{remove_from_log}
\title{Delete from log}
\usage{
remove_from_log(.tbl, log, id_col = "uuid")
}
\arguments{
\item{.tbl}{A tibble}

\item{log}{A deletion log, which contains a column "action".}

\item{id_col}{The column id, usually "uuid"}
}
\value{
A lot of removal
}
\description{
Removes surveys from .tbl where column "action" in the log is set to "remove".
}
