% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rename_cols}
\alias{rename_cols}
\title{Rename columns}
\usage{
rename_cols(.tbl, old_cols, new_cols)
}
\arguments{
\item{.tbl}{A tible}

\item{old_cols}{A vector of old quoted variable names}

\item{new_cols}{A vector of new quoted variable names}
}
\value{
Updated tibble with new names
}
\description{
Rename columns
}
