% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{rowwise_optimum}
\alias{rowwise_optimum}
\title{Tidy rowwise optimum (default to pmax) of several columns}
\usage{
rowwise_optimum(
  .tbl,
  ...,
  optimum = "max",
  max_name = "pmax",
  min_name = "pmin",
  na_rm = TRUE,
  keep = "all"
)
}
\arguments{
\item{.tbl}{A tibble}

\item{...}{Unquoted numeric column names}

\item{optimum}{Should we calculate "max", "min" or "both"? Default to TRUE}

\item{max_name}{Column name for the mutated pmax}

\item{min_name}{Column name for the mutated pmin}

\item{na_rm}{Remove NAs. Default to TRUE}

\item{keep}{To be used within mutate. Default to "all".}
}
\value{
A tibble with pmax, pmin or both (and all columns or none, depending on "keep")
}
\description{
Tidy rowwise optimum (default to pmax) of several columns
}
