% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_f.R
\name{scale_fill_reach}
\alias{scale_fill_reach}
\title{Fill scale constructor for reach colors}
\usage{
scale_fill_reach(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in drsimonj_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_fill_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\value{
A fill scale for ggplot
}
\description{
Fill scale constructor for reach colors
}
