% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_alm.R
\name{score_alm}
\alias{score_alm}
\title{Extract overall almetric score for a publication}
\usage{
score_alm(id_list)
}
\arguments{
\item{id_list}{Vector of unique Digital Object Identifiers (DOI) and/or PMID}
}
\value{
Vector of overall almetric scores (or NA if not available)
}
\description{
Extract overall almetric score for a publication
}
