% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_hindex.R
\name{score_hindex}
\alias{score_hindex}
\title{Calculate an H-Index based on number of citations.}
\usage{
score_hindex(n_citations)
}
\arguments{
\item{n_citations}{Vector of citations (numerical variable).}
}
\value{
H-Index score
}
\description{
Calculate an H-Index based on a vector of citations.
}
