% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_pubmed.R
\name{search_pubmed}
\alias{search_pubmed}
\title{Perform a focused pubmed search to identify the PMID associated with specified authors or registry identifiers.}
\usage{
search_pubmed(
  search_type = "author",
  search_list,
  date_min = NULL,
  date_max = Sys.Date(),
  keywords = NULL
)
}
\arguments{
\item{search_type}{Type of search desired (either "author" or "registry_id").}

\item{search_list}{Vector of search terms (either author last name and initial, or registry identifers).}

\item{date_min}{String of a date in "Year-Month-Day" format to provide  a lower limit to search within (default = NULL).}

\item{date_max}{String of a date in "Year-Month-Day" format to provide an upper limit to search within (default = current date).}

\item{keywords}{Vector of keywords or patterns that are required to be present for inclusion (default = NULL).}
}
\value{
Vector of pubmed identifiers (PMID)
}
\description{
Search pubmed to identify PMID associated with specified authors or registry identifiers.
}
