% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{severity_score_reach}
\alias{severity_score_reach}
\title{MSNA severity scores names and colors}
\usage{
severity_score_reach()
}
\value{
A tibble with three columns: color, level, name of severity scores
}
\description{
MSNA severity scores names and colors
}
