% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sift.R
\name{sift}
\alias{sift}
\title{Sift though pubmed records to stratify by potential relevance}
\usage{
sift(
  extract,
  authors = NULL,
  affiliations = NULL,
  keyword = NULL,
  ignoreword = NULL,
  separate = T
)
}
\arguments{
\item{authors}{Vector of last name and initial(s) which correspond to relevant authors (default = NULL).}

\item{affiliations}{Vector of strings which correspond to relevant author affiliations (default = NULL).}

\item{keyword}{Vector of keyword or patterns that are expected to be present in the title/abstract/journal (default = NULL).}

\item{ignoreword}{Vector of keyword or patterns that are expected to NOT be present in the title/abstract/journal (default = NULL).}

\item{separate}{Logical value to separate the output into wheat (highlighted publications) and chaff (unhighlighted publications).}

\item{pmid}{Vector of pubmed identifiers (PMID) desired to be screened.}
}
\value{
Tibble containing pmid, results from sifting based on the authors, affiliations, and keyword, and the title/journal/abstract to allow further evaluation.
}
\description{
Highlight pubmed records potentially relevant to the authors of interest
}
