% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sift_pubmed.R
\name{sift_pubmed}
\alias{sift_pubmed}
\title{Sift though pubmed records to stratify by potential relevance}
\usage{
sift_pubmed(
  pmid,
  authors = NULL,
  affiliations = NULL,
  keywords = NULL,
  separate = T
)
}
\arguments{
\item{pmid}{Vector of pubmed identifiers (PMID) desired to be screened.}

\item{authors}{Vector of last name and initial(s) which correspond to relevant authors (default = NULL).}

\item{affiliations}{Vector of strings which correspond to relevant author affiliations (default = NULL).}

\item{keywords}{Vector of keywords or patterns that are expected to be present in the title/abstract/journal (default = NULL).}

\item{separate}{Logical value to separate the output into wheat (highlighted publications) and chaff (unhighlighted publications).}
}
\value{
Tibble containing pmid, results from sifting based on the authors, affiliations, and keywords, and the title/journal/abstract to allow further evaluation.
}
\description{
Highlight pubmed records potentially relevant to the authors of interest
}
