% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_kobo.R
\name{split_survey}
\alias{split_survey}
\title{Split survey type and list name}
\usage{
split_survey(
  survey,
  col_to_split,
  into = c("type", "list_name"),
  sep = " ",
  fill = "right",
  ...
)
}
\arguments{
\item{survey}{A survey sheet from Kobo}

\item{col_to_split}{Usually `type`}

\item{into}{Vector of columns names to split to. Default to c("type", "list_name" )}

\item{sep}{The separator. Default to " ".}

\item{fill}{How to fill. Default to NA on the right.}

\item{...}{Params to pass to `tidyr::separate`}
}
\value{
A survey tibble, split
}
\description{
Split survey type and list name
}
