% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_df.R
\name{string_count}
\alias{string_count}
\title{Count the number of occurrences of a string over a data frame}
\usage{
string_count(df, pattern = "", new_colname = "count")
}
\arguments{
\item{df}{A dataframe}

\item{pattern}{A pattern to pass to `stringr::str_count()`. Default to "".}

\item{new_colname}{The newly-mutated colname. Default to "count".}
}
\value{
A mutated dataframe
}
\description{
Count the number of occurrences of a string over a data frame
}
