% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_vec.R
\name{sublist_start}
\alias{sublist_start}
\title{Sublist of start pattern of list names}
\usage{
sublist_start(list, pattern_start)
}
\arguments{
\item{list}{A list}

\item{pattern_start}{A pattern}
}
\value{
A sublist
}
\description{
Sublist of start pattern of list names
}
