% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_vec.R
\name{subvec_in}
\alias{subvec_in}
\title{Subvec in}
\usage{
subvec_in(vector, set)
}
\arguments{
\item{vector}{A vector to subset}

\item{set}{A set-vector}
}
\value{
A subset of a list or a vector
}
\description{
Subvec in
}
