% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subvec_not_in}
\alias{subvec_not_in}
\title{Subvec not in}
\usage{
subvec_not_in(vector, set)
}
\arguments{
\item{vector}{A vector to subset}

\item{set}{A set-vector}
}
\value{
A subset of vector not in set
}
\description{
Subvec not in
}
