% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_df.R
\name{survey_duration}
\alias{survey_duration}
\title{Calculate the duration of a survey}
\usage{
survey_duration(.tbl, start, end, new_colname = "survey_duration")
}
\arguments{
\item{.tbl}{A tibble}

\item{start}{Start column name}

\item{end}{End column name}

\item{new_colname}{The new column name of the time duration}
}
\value{
A tibble with three new colums, including the duration of survey in minutes
}
\description{
Calculate the duration of a survey
}
\details{
Note: it is necessary to have 'start' and 'end' columns
}
