% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{svy_diff_time}
\alias{svy_diff_time}
\title{Check in-between surveys time}
\usage{
svy_diff_time(.tbl, id_enum, start, end)
}
\arguments{
\item{.tbl}{A tibble}

\item{id_enum}{The enumerator id column}

\item{start}{Start column (unquoted)}

\item{end}{End column (unquoted)}
}
\value{
A tibble with ... removed
}
\description{
Check in-between surveys time
}
\details{
Note: it is necessary to have 'start' and 'end' columns with no NA
}
