% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{svy_interact}
\alias{svy_interact}
\title{Survey interaction means}
\usage{
svy_interact(design, interact_cols, group = NULL, unnest_interaction = T, ...)
}
\arguments{
\item{design}{A srvyr::design object}

\item{interact_cols}{A vector of columns to get interactions from}

\item{group}{A vector of columns to group by. This is the default.}

\item{unnest_interaction}{Should interaction be unnested? Default to TRUE.}

\item{...}{Parameters to pass to srvyr::survey_mean()}
}
\value{
A survey-summarized-interaction tibble
}
\description{
Survey interaction means
}
