% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{svy_median}
\alias{svy_median}
\title{Survey median}
\usage{
svy_median(design, col, group = NULL, na_rm = T, stat_name = "median", ...)
}
\arguments{
\item{design}{A srvyr::design object}

\item{col}{A column to calculate median from}

\item{group}{A quoted or unquoted vector of columns to group by. Default to NULL for no group}

\item{na_rm}{Should NAs from `col` be removed? Default to TRUE. na.rm does not work anymore within srvyr functions (workaround for now). It should work for `survey_median`, matter of precaution}

\item{stat_name}{What should the statistic's column be named? Default to "median"}

\item{...}{Parameters to pass to `srvyr::survey_median()`}
}
\value{
A survey-summarized-proportion tibble
}
\description{
Survey median
}
