% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{svy_prop}
\alias{svy_prop}
\title{Survey prop}
\usage{
svy_prop(design, col, group = NULL, ...)
}
\arguments{
\item{design}{A srvyr::design object}

\item{col}{A column to calculate proportion from}

\item{group}{A quoted or unquoted vector of columns to group by. Default to NULL for no group.}

\item{...}{Parameters to pass to `srvyr::survey_prop()`}
}
\value{
A survey-summarized-proportion tibble
}
\description{
Survey prop
}
