% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tbl_col_end}
\alias{tbl_col_end}
\title{Get colnames that end with pattern}
\usage{
tbl_col_end(.tbl, pattern_end)
}
\arguments{
\item{.tbl}{A tibble}

\item{pattern_end}{A pattern}
}
\value{
A vector of column names
}
\description{
Get colnames that end with pattern
}
