% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_vec.R
\name{tbl_col_start}
\alias{tbl_col_start}
\title{Get colnames that start with pattern}
\usage{
tbl_col_start(.tbl, pattern_start)
}
\arguments{
\item{.tbl}{A tibble}

\item{pattern_start}{A pattern}
}
\value{
A vector of column names
}
\description{
Get colnames that start with pattern
}
