% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{theme_flip_simple_reach}
\alias{theme_flip_simple_reach}
\title{Some reach more minimal theme for ggplot}
\usage{
theme_flip_simple_reach(family = "Leelawadee UI")
}
\arguments{
\item{family}{A character name of the font family. Default to "Leelawadee UI".}
}
\value{
A theme to be added to the "+" ggplot grammar for `impactR::reach_flip_bar()`
}
\description{
Give some reach colors and fonts to a ggplot. Based on theme_bw().
}
