% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{theme_hist_simple_reach}
\alias{theme_hist_simple_reach}
\title{Some reach more minimal theme for ggplot}
\usage{
theme_hist_simple_reach(family = "Leelawadee UI")
}
\arguments{
\item{family}{A character name of the font family. Default to "Leelawadee UI".}
}
\value{
A theme to be added to the "+" ggplot grammar for `impactR::reach_hist_bar()`
}
\description{
Give some reach colors and fonts to a ggplot. Based on theme_bw().
}
