% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{theme_reach}
\alias{theme_reach}
\title{Some reach theme for ggplot}
\usage{
theme_reach(family = "Leelawadee UI")
}
\arguments{
\item{family}{A character name of the font family. Default to "Leelawadee UI".}
}
\value{
A theme to be added to the "+" ggplot grammar
}
\description{
Give some reach colors and fonts to a ggplot. Based on theme_bw().
}
\details{
##########################
##########################
}
