% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{tmap_options_reach}
\alias{tmap_options_reach}
\title{Some  REACH basic `tmap` options}
\usage{
tmap_options_reach(
  check_and_fix = T,
  fontfamily = "Leelawadee UI",
  fontface = 1,
  legend_title_size = 0.9,
  legend_text_size = 0.8,
  legend_text_fontface = 1,
  legend_width = 0.9,
  legend_position = c("right", 0.14),
  legend_frame = impactR::reach_colors()$reach_dk_grey,
  output_dpi = 500,
  frame = impactR::reach_cols("main_grey"),
  inner_margins = c(0, 0, 0, 0),
  outer_margins = c(0, 0, 0, 0),
  asp = 0,
  ...
)
}
\arguments{
\item{check_and_fix}{Param check.and.fix from `tmap::tmap_options()`}

\item{fontfamily}{Param fontfamily from `tmap::tmap_options()`}

\item{fontface}{Param fontface from `tmap::tmap_options()`}

\item{legend_title_size}{Param legend.title.size from `tmap::tmap_options()`}

\item{legend_text_size}{Param legend.text.size from `tmap::tmap_options()`}

\item{legend_text_fontface}{Param legend legend.text.fontface from `tmap::tmap_options()`}

\item{legend_width}{Param legend.width from `tmap::tmap_options()`}

\item{legend_position}{Param legend.position from `tmap::tmap_options()`}

\item{legend_frame}{Param legend.frame from `tmap::tmap_options()`}

\item{output_dpi}{Param output.dpi from `tmap::tmap_options()`}

\item{frame}{Frame colors using `impactR::reach_cols()`}

\item{inner_margins}{Param inner.margins from `tmap::tmap_options()`}

\item{outer_margins}{Param outer.margins from `tmap::tmap_options()`}

\item{asp}{Param asp from `tmap::tmap_options()`}

\item{...}{Other parameters to pass to `tmap::tmap_options()`}
}
\value{
Load tmap options globally
}
\description{
Give some reach colors and font for maps
}
