% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{update_rows_from_list}
\alias{update_rows_from_list}
\title{Update rows from list}
\usage{
update_rows_from_list(.tbl, .list, id_col = "uuid")
}
\arguments{
\item{.tbl}{A tibble}

\item{.list}{A list of tibbles with values to update}

\item{id_col}{The column id, usually "uuid"}
}
\value{
A tibble with values updated
}
\description{
Update rows from list
}
