% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cleaning_log}
\alias{cleaning_log}
\title{Some random cleaning log}
\format{
A data frame with 17 rows and 17 variables:
\describe{
\item{id_check}{The check id}
\item{name}{Tne choices's names}
...
}
}
\usage{
cleaning_log
}
\description{
A dataset (in French) containing a cleaning_log. Columns are necessary and comes from the \code{make_log()} function
}
\keyword{datasets}
