% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dap}
\alias{dap}
\title{Some random dap test}
\format{
A data frame with 5 rows and 13 variables:
\describe{
\item{id_analysis}{The analysis id. Each line is a different one.}
\item{rq}{The topic/sector of the indicator.}
\item{sub_rq}{The sub-topic/sub-sector of the indicator.}
\item{indicator}{The label of the indicator, e.g. "\\% of HHs by type of shelter".}
\item{recall}{The recall period (not to be included in the indicator), e.g. "30 days prior to data collection".}
\item{question_name}{The name of the variable in the data or, for a ratio, the two names separated by a comma (no space).}
\item{subset}{If calculated on a subset, name it, e.g. "among households that received a humanitarian aid".}
\item{analysis_name}{The analysis type label, e.g. "Percentage" or "Mean".}
\item{analysis}{The R type of analysis to pass to \code{make_analysis()}.}
\item{none_label}{The label for all NA values if "prop_simple_overall" or "prop_multiple_overall", e.g. "None or missing data".}
\item{group_name}{The label of the grouping variable/column, e.g. "Household displacement status" with codes that may be labelled "IDP" or "Non IDP".}
\item{group}{The variable/column name to group by, e.g. "hh_status" that may be coded "idp" or non_idp".}
\item{level}{Level of confidence. Parameter to pass to \verb{svy_*()} functions.}
\item{na_rm}{Should NAs be removed? Parameter to pass to \verb{svy_*()} functions.}
\item{vartype}{Variance type in outputs. Parameter to pass to \verb{svy_*()} functions.}
}
}
\usage{
dap
}
\description{
A dataset (in French) containing a basic dap, just to test. Very experimental
}
\keyword{datasets}
