% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{define_region}
\alias{define_region}
\title{Define region of interest}
\usage{
define_region(data, start_time, end_time)
}
\arguments{
\item{data}{An \code{impactr_data} object, as obtained with
\link[=read_acc]{read_acc()}.}

\item{start_time, end_time}{A character string with the start and end times
of the region of interest in the "YYYY-MM-DD HH:MM:SS" format.}
}
\value{
An object of class \code{impactr_data}.
}
\description{
Define the region of interest for data analysis based on the accelerometer
data timestamp.
}
\examples{
data <- read_acc(impactr_example("hip-raw.csv"))
define_region(
  data, start_time = "2021-04-06 15:45:00", end_time = "2021-04-06 15:46:00"
)
}
