% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators_foodsec.R
\name{fcm}
\alias{fcm}
\title{FCM - FEWS NET Food Consumption Matrix}
\usage{
fcm(.tbl, fcs_cat, hhs_cat, rcsi_cat)
}
\arguments{
\item{.tbl}{Data}

\item{fcs_cat}{Column produced by `impactR::fcs()`}

\item{hhs_cat}{Column produced by `impactR::hhs()`}

\item{rcsi_cat}{Column produced by `impactR::rcsi()`}
}
\value{
The FEWS NET Food Consumption Cell number.
}
\description{
FCM - FEWS NET Food Consumption Matrix
}
