% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_utils.R
\name{filter_signal}
\alias{filter_signal}
\title{Filter a signal}
\usage{
filter_signal(...)
}
\arguments{
\item{...}{Arguments to be passed to the python function:
\itemize{
\item x: A numeric vector.
\item n: The order of the filter.
\item wn: The critical frequency or frequencies.
\item type: The type of the filter.
}}
}
\value{
A numeric vector with the filtered signal.
}
\description{
Function imported from python. Apply a butterworth digital filter to a
signal. For more information run
\code{\link[reticulate:py_help]{reticulate::py_help(filter_signal)}}.
}
