% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_intersect.R
\name{format_intersect}
\alias{format_intersect}
\title{Additional function for ComplexHeatmap}
\usage{
format_intersect(data_upset)
}
\arguments{
\item{comb_mat}{Combination matrix output from ComplexHeatmap::make_comb_mat()}
}
\value{
Tibble of combinations with the name of the sets provided.
}
\description{
Additional function for ComplexHeatmap to faciltate named sets in combinations
}
