% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_multiple}
\alias{get_multiple}
\title{Get all select multiple from survey sheet}
\usage{
get_multiple(survey)
}
\arguments{
\item{survey}{The survey sheet from Kobo (with column "type" split)}
}
\value{
A character vector of select_multiple questions
}
\description{
Get all select multiple from survey sheet
}
\details{
survey should have a split type column with types of variables such as "select_one", "select_multiple", etc.
}
