% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_cite.R
\name{impact_cite}
\alias{impact_cite}
\title{Extract additional traditional metrics from Google Scholar}
\usage{
impact_cite(
  df,
  id,
  crossref = TRUE,
  dimentions = TRUE,
  scopus = FALSE,
  oc = FALSE,
  gscholar = FALSE,
  gscholar_title_nchar = 50,
  metric = TRUE
)
}
\arguments{
\item{df}{Dataframe containing at least two columns: publication year ("year") and id with each publication listed as a row.}

\item{id}{String of the column name containing the publication id (accepts DOI and PMID).}

\item{crossref}{Crossref database to be used for citation data (default=TRUE)}

\item{dimentions}{Dimentions database to be used for citation data (default=TRUE)}

\item{scopus}{Scopus database to be used for citation data (default=FALSE). Requires Scopus API.}

\item{oc}{Open citations database to be used for citation data (default=FALSE due to sparse population at present).}

\item{gscholar}{Google scholar database to be used for citation data (default=FALSE). Requires Google Scholar ID (and all relevent publications to be associated with this).}

\item{gscholar_title_nchar}{Argument to specify how many characters the titles should be matched (default = 50). See "gscholar_invalid" output.}

\item{metric}{Return common citation metrics (default = TRUE)}
}
\value{
Nested dataframe of: (1)."df": Amended dataframe with additional citation data appended (2). "time": Dataframe of citations over time (only avaiable for Scopus and google scholar). (3). "metric": output from cite_metric() (4). Unmatched recorded (only for google scholar)
}
\description{
Extract additional traditional metrics from Google Scholar
}
