% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{make_analysis_from_dap}
\alias{make_analysis_from_dap}
\title{Make analysis from data analysis plan}
\usage{
make_analysis_from_dap(design, survey, choices, dap)
}
\arguments{
\item{design}{A design object}

\item{survey}{The survey sheet from Kobo that contains at least column 'list_name' (split from 'type') and 'name'}

\item{choices}{The choices sheet from Kobo contains at least column 'list_name' (split from 'type') and 'name'}

\item{dap}{A data analysis plan, typically from an excel sheet}
}
\value{
A summarized analysis
}
\description{
Make analysis from data analysis plan
}
\concept{functions for survey analysis

#' @description
`r lifecycle::badge("experimental")`

This function still is experimental.}
