% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{make_log_from_check_list}
\alias{make_log_from_check_list}
\title{Apply `make_log` to a tibble of check_list (one logical test per row)}
\usage{
make_log_from_check_list(
  .tbl,
  survey,
  check_list,
  id_col = "uuid",
  cols_to_keep
)
}
\arguments{
\item{.tbl}{A tibble of data}

\item{survey}{The survey sheet of the Kobo tool}

\item{check_list}{A tibble of logical test to check for}

\item{id_col}{Survey id, usually "uuid"}

\item{cols_to_keep}{Columns to keep in the log, e.g. today, enum_id}
}
\description{
Apply `make_log` to a tibble of check_list (one logical test per row)
}
\seealso{
[make_log] for a log based on a logical test

[make_log_outlier()] for a log based on outliers

[make_log_other()] for a log of others

[make_all_logs()] for all logs: outliers, others, logical tests

Other functions to make logs: 
\code{\link{make_all_logs}()},
\code{\link{make_log_other}()},
\code{\link{make_log_outlier}()},
\code{\link{make_log}()}
}
\concept{functions to make logs}
