% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{modify_from_log}
\alias{modify_from_log}
\title{Modify from log}
\usage{
modify_from_log(.tbl, log, id_col, type, other)
}
\arguments{
\item{.tbl}{A tibble}

\item{log}{A cleaning log}

\item{id_col}{The column id, usually "uuid"}

\item{type}{Either "character" or "double"}

\item{other}{A character vector of the start of all other column names. E.g., other = "other_"}
}
\value{
A tibble with values modified
}
\description{
Modify from log
}
