% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{reach_cols}
\alias{reach_cols}
\title{Function to extract reach colors as hex codes}
\usage{
reach_cols(..., unnamed = T)
}
\arguments{
\item{...}{Character names of reach colors. If NULL returns all colors}

\item{unnamed}{Should the output vector be unnamed? Default to T}
}
\value{
An hex code
}
\description{
Function to extract reach colors as hex codes
}
\details{
This function needs to be modified to add colors
}
