% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_acc.R
\name{read_acc}
\alias{read_acc}
\title{Read raw accelerometer data}
\usage{
read_acc(file)
}
\arguments{
\item{file}{Path to a raw accelerometer data file.}
}
\value{
An object of class \code{impactr_data}.
}
\description{
Reads raw accelerometer data files into an \code{impactr_data} object.
}
\examples{
read_acc(impactr_example("hip-raw.csv"))
}
