% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rec_na}
\alias{rec_na}
\title{Replace many types of NAs of a tibble to NA}
\usage{
rec_na(.tbl, ...)
}
\arguments{
\item{.tbl}{A tibble}

\item{...}{Column names}
}
\value{
A tibble with one type of NAs
}
\description{
Replace many types of NAs of a tibble to NA
}
