% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rm_cols}
\alias{rm_cols}
\title{Remove cols from tibble}
\usage{
rm_cols(.tbl, cols)
}
\arguments{
\item{.tbl}{A tibble}

\item{cols}{A vector of column names (quoted)}
}
\value{
A tibble with columns removed
}
\description{
Remove cols from tibble
}
