% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{svy_difftime}
\alias{svy_difftime}
\title{Check in-between surveys time}
\usage{
svy_difftime(.tbl, start, end, ..., new_colname = "survey_difftime")
}
\arguments{
\item{.tbl}{A tibble}

\item{start}{Start column (unquoted)}

\item{end}{End column (unquoted)}

\item{...}{The columns to group by, usually a locality or an enumerator id}

\item{new_colname}{The new column name of the time difference}
}
\value{
A tibble with ... removed
}
\description{
Check in-between surveys time
}
\details{
Note: it is necessary to have 'start' and 'end' columns with no NA
}
