% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{svy_duration}
\alias{svy_duration}
\title{Calculate the duration of a survey}
\usage{
svy_duration(.tbl, start = "start", end = "end")
}
\arguments{
\item{.tbl}{A tibble}

\item{start}{Start column name. Default to start}

\item{end}{End column name. Default to end}
}
\value{
A tibble with three new colums, including the duration of survey in minutes
}
\description{
Calculate the duration of a survey
}
\details{
Note: it is necessary to have 'start' and 'end' columns
}
