% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{svy_mean}
\alias{svy_mean}
\title{Survey mean}
\usage{
svy_mean(design, col, group = NULL, na_rm = T, ...)
}
\arguments{
\item{design}{A srvyr::design object}

\item{col}{A column to calculate mean from}

\item{group}{A quoted or unquoted vector of columns to group by. Default to NULL for no group.}

\item{na_rm}{Should NAs from `col` be removed? Default to TRUE. na.rm does not work anymore within srvyr functions (workaround for now). It should work for `survey_mean`, matter of precaution}

<<<<<<< Updated upstream
=======
\item{stat_name}{What should the statistic's column be named? Default to "mean"}

>>>>>>> Stashed changes
\item{...}{Parameters to pass to `srvyr::survey_mean()`}
}
\value{
A survey-summarized-proportion tibble
}
\description{
Survey mean
}
