% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{svy_prop}
\alias{svy_prop}
\title{Survey proportion}
\usage{
svy_prop(design, col, group = NULL, na_rm = T, stat_name = "prop", ...)
}
\arguments{
\item{design}{A srvyr::design object}

\item{col}{A column to calculate proportion from}

\item{group}{A quoted or unquoted vector of columns to group by. Default to NULL for no group.}

\item{na_rm}{Should NAs from `col` be removed? Default to TRUE. na.rm does not work anymore within srvyr functions (workaround for now)}

\item{stat_name}{What should the statistic's column be named? Default to "ratio"}

\item{...}{Parameters to pass to `srvyr::survey_prop()`}
}
\value{
A survey-summarized-proportion tibble
}
\description{
Survey proportion
}
