% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{svy_ratio}
\alias{svy_ratio}
\title{Survey ratio}
\usage{
svy_ratio(
  design,
  num,
  denom,
  group = NULL,
  na_rm = T,
  stat_name = "ratio",
  ...
)
}
\arguments{
\item{design}{A srvyr::design object}

\item{num}{The numerator column}

\item{denom}{The denominator column}

\item{group}{A quoted or unquoted vector of columns to group by. Default to NULL for no group}

\item{na_rm}{Should NAs from \code{num} and \code{denom} be removed? Default to TRUE. na.rm does not work anymore within srvyr functions (workaround for now). It should work for \code{survey_mean}, matter of precaution}

\item{stat_name}{What should the statistic's column be named? Default to "ratio"}

\item{...}{Parameters to pass to srvyr::survey_mean()}
}
\value{
A survey-summarized-interaction tibble
}
\description{
Survey ratio
}
